Fixing incorrect calculation in hilbert ranges
If we have a integer domain from -1 -> 2 with 2 bits hilbert values,
we will have a normalized domain of 0, 1, 2, 3 as well as a bucket_num
value of 3. Old behavior had a dom_range_T value of 4, which turns
every possible value to 0, 0.75, 1.5, 2.25 which would end up rounded
to 0, 0, 1, 2. With the correct dom_range_T of 3, we end up with the
correct Hilbert values of 0, 1, 2, 3.

---
TYPE: BUG
DESC: Fixing incorrect calculation in hilbert ranges